--- src/layouts/Default/index.js
@@ -1,33 +1,0 @@
-import React from "react"
-import PropTypes from "prop-types"
-
-import GlobalStyle from "@layouts/global"
-import Header from "@components/Header"
-
-const Layout = ({ children }) => (
-  <>
-    <GlobalStyle />
-    <Header />
-    
-    <div
-      style={{
-        margin: `0 auto`,
-        maxWidth: 960,
-        padding: `0 1.0875rem 1.45rem`,
-      }}
-    >
-      <main>{children}</main>
-      <footer>
-        © {new Date().getFullYear()}, Built with
-        {` `}
-        <a href="https://www.gatsbyjs.org">Gatsby</a>
-      </footer>
-    </div>
-  </>
-)
-
-Layout.propTypes = {
-  children: PropTypes.node.isRequired,
-}
-
-export default Layout
