+++ src/layouts/Default/theme.js
@@ -1,0 +1,10 @@
+const theme = {
+  colors: {
+    dark: '#000',
+    white: '#fff',
+    primary: '#47bb8e',
+  },
+  transition: 'all 0.3s ease-in-out',
+}
+
+export default theme
