+++ src/layouts/media.js
@@ -10,13 +10,16 @@ const sizes = {
     large: '1169',
     medium: '767',
     small: '449',
-  }
+  },
 }
 
-const media = Object.keys(sizes.min).reduce((acc, label) => {
-  acc.min[label] = `(min-width: ${sizes.min[label]}px)`
-  acc.max[label] = `(max-width: ${sizes.max[label]}px)`
-  return acc
-}, { min: {}, max: {} })
+const media = Object.keys(sizes.min).reduce(
+  (acc, label) => {
+    acc.min[label] = `(min-width: ${sizes.min[label]}px)`
+    acc.max[label] = `(max-width: ${sizes.max[label]}px)`
+    return acc
+  },
+  { min: {}, max: {} }
+)
 
 export default media
